+++ components/adobe-acrobat-sign/src/actions/users/listUsers.ts
@@ -10,7 +10,7 @@ export const listUsers = action({
     description: "Retrieves all the users in an account.",
   },
   inputs: listUsersInputs,
-  perform: async (context, { connection, fetchAll, cursor, pageSize }) => {
+  perform: async (context, { connection, fetchAll, pagination }) => {
     const client = getAdobeSignClient(connection, context.debug.enabled);
     const data = await fetchAdobeSignResults<
       UserResponse,
@@ -21,8 +21,8 @@ export const listUsers = action({
       "/users",
       fetchAll,
       {
-        pageSize: pageSize || undefined,
-        cursor: cursor || undefined,
+        pageSize: pagination.pageSize || undefined,
+        cursor: pagination.cursor || undefined,
       },
       "userInfoList",
     );